+++ Classes/Configuration/CsvExportConfigurationCheck.php
@@ -19,7 +19,6 @@ class CsvExportConfigurationCheck extends AbstractConfigurationCheck
         $this->checkFieldsFromAttendanceForCsv();
         $this->checkFieldsFromFeUserForEmailCsv();
         $this->checkFieldsFromAttendanceForEmailCsv();
-        $this->checkShowAttendancesOnRegistrationQueueInEmailCsv();
     }
 
     private function checkFieldsFromFeUserForCsv(): void
@@ -60,16 +59,6 @@ class CsvExportConfigurationCheck extends AbstractConfigurationCheck
             'These values specify the registration fields to export via CSV in email mode.
             A mistyped field name will cause the field to not get included.',
             'tx_seminars_attendances'
-        );
-    }
-
-    private function checkShowAttendancesOnRegistrationQueueInEmailCsv(): void
-    {
-        $this->checkIfBoolean(
-            'showAttendancesOnRegistrationQueueInEmailCsv',
-            'This value specifies if attendances on the registration queue should also be exported in the CSV file
-            in the email mode.
-            If this is not set correctly, the attendances on the registration queue might not get exported.'
         );
     }
 }