+++ Tests/LegacyFunctional/Csv/EmailRegistrationListViewTest.php
@@ -223,11 +223,8 @@ final class EmailRegistrationListViewTest extends FunctionalTestCase
     /**
      * @test
      */
-    public function renderForQueueRegistrationsNotAllowedForEmailNotContainsRegistrationOnQueue(): void
+    public function renderForQueueRegistrationsNotContainsRegistrationOnQueue(): void
     {
-        $this->configuration->setAsBoolean('showAttendancesOnRegistrationQueueInEmailCsv', false);
-        $this->configuration->setAsBoolean('showAttendancesOnRegistrationQueueInCSV', true);
-
         $this->configuration->setAsString('fieldsFromAttendanceForCsv', 'uid');
         $this->configuration->setAsString('fieldsFromAttendanceForEmailCsv', 'uid');
 
@@ -243,34 +240,6 @@ final class EmailRegistrationListViewTest extends FunctionalTestCase
         );
 
         self::assertStringNotContainsString(
-            (string)$registrationUid,
-            $this->subject->render()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function renderForQueueRegistrationsAllowedForEmailNotContainsRegistrationOnQueue(): void
-    {
-        $this->configuration->setAsBoolean('showAttendancesOnRegistrationQueueInEmailCsv', true);
-        $this->configuration->setAsBoolean('showAttendancesOnRegistrationQueueInCSV', false);
-
-        $this->configuration->setAsString('fieldsFromAttendanceForCsv', 'uid');
-        $this->configuration->setAsString('fieldsFromAttendanceForEmailCsv', 'uid');
-
-        $registrationUid = $this->testingFramework->createRecord(
-            'tx_seminars_attendances',
-            [
-                'seminar' => $this->eventUid,
-                'crdate' => (int)GeneralUtility::makeInstance(Context::class)
-                    ->getPropertyFromAspect('date', 'timestamp'),
-                'user' => $this->testingFramework->createFrontEndUser(),
-                'registration_queue' => true,
-            ]
-        );
-
-        self::assertStringContainsString(
             (string)$registrationUid,
             $this->subject->render()
         );